+++ src/zeusops_bot/discord.py
@@ -5,6 +5,7 @@ from logging import Logger
 from discord import Bot
 
 from zeusops_bot.cogs import ZeusUpload
+from zeusops_bot.reforger_config_gen import ReforgerConfigGenerator
 from zeusops_bot.settings import ZeusopsBotConfig
 
 
@@ -16,8 +17,12 @@ class ZeusopsBot(Bot):
         super().__init__(*args, **kwargs)
         self.config = config
         self.logger = logger
+        self.reforger_confgen = ReforgerConfigGenerator(
+            base_config_file=config.reforger.reference_config,
+            target_folder=config.reforger.config_folder,
+        )
 
-        cog = ZeusUpload(self, config)
+        cog = ZeusUpload(self, self.reforger_confgen)
         self.add_cog(cog)
 
     def run(self):